Add direct upload support to GCS service
Google Cloud Storage does not ask you to (or allow?) specify the content
length up front, so that parameter is ignored right now. Otherwise, the
functionality is similar to the S3 service's direct upload capability.

The integration test for uploading the written URL was failing similarly
to David's S3 test when using HTTParty. Dropping down to use Net::HTTP
allows the test to complete successfully by writing and then downloading
the resulting file.